src/node/text_formatting.rs: revising `format_text`
