+++ src/node/node_ref.rs
@@ -21,7 +21,7 @@ use super::child_nodes;
 use super::id_provider::NodeIdProver;
 use super::inner::TreeNode;
 use super::node_data::NodeData;
-use super::serializing::format_text;
+use super::text_formatting::format_text;
 use super::serializing::SerializableNodeRef;
 use super::NodeId;
